sign in flow for accounts menu
